Templates attached to Django
sign up and sign in template is integrated with django and all the details will be added to database according and login's can be done.
New error boxes have been added to the Login html file.
But a little styling is required to be done for the error boxes which was added in templates